frontend-library/src/service_side/python_part/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
pandas==1.3.3
requests==2.25.1
scikit-surprise==1.1.4
gunicorn==20.1.0
python-dotenv==1.0.0
